ia/AzureOpenAI/Assistants/requirements.txt
openai
IPython
python-dotenv